+++ breaking-front/src/pages/ProfileEdit/hooks/queries/useProfileDetailData.js
